Rename table name pattern to identifier pattern
- Rename TableNamePatternType to IdentifierNamePatternType
- Update related methods and classes to use the new IdentifierNamePatternType
- Modify the formatTableNamePattern method to formatIdentifierPattern
